Move piece-related data structure into a separate module
